requirements.txt
Flask==2.1.2
requests==2.26.0
fpdf==1.7.2
python-dotenv==0.19.2
